feat: update sign in and sign up page
